requirements.txt
pandas>=2.0.0
openpyxl>=3.0.10
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
langchain>=0.0.300
langgraph>=0.6.3
langchain-core>=0.0.1
google-generativeai>=0.3.0    
pydantic>=1.10.0
typing-extensions>=4.0.0
dateparser>=1.1.1
streamlit>=1.25.0
filelock>=3.12.0
requests>=2.31.0
python-dateutil>=2.8.2
xlsxwriter==3.1.9